requirements.txt
Django
requests
lxml
python-dateutil